requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic-settings==2.2.1